Start splitting off stuff into NamespaceDiffs2
