+++ package.yaml
@@ -109,6 +109,7 @@ dependencies:
 - share-auth
 - unison-hashing-v2
 - unison-codebase-sqlite-hashing-v2
+- unison-merge
 - unison-parser-typechecker
 - unison-prelude
 - unison-pretty-printer
