+++ stack.yaml
@@ -31,6 +31,7 @@ packages:
 - unison/parser-typechecker
 - unison/unison-core
 - unison/unison-hashing-v2
+- unison/unison-merge
 - unison/unison-share-api
 - unison/unison-share-projects-api
 - unison/unison-syntax
